+++ anvil_consortium_manager/models.py
@@ -652,7 +652,11 @@ class Workspace(TimeStampedModel):
         help_text="Indicator of whether the workspace is locked or not.",
         default=False,
     )
-
+    is_requester_pays = models.BooleanField(
+        verbose_name="Requester pays",
+        help_text="Indicator of whether the workspace is set to requester pays.",
+        default=False,
+    )
     history = HistoricalRecords()
 
     class Meta: